// src/interfaces/model/accessRightsObject.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AllowedUrlObject } from './allowedUrlObject'
import { LimitObject } from './limitObject'
import { QuotaObject } from './quotaObject'

/**
 * Access rights data
 */
export interface AccessRightsObject {
  /**
   * Available urls and methods list
   */
  allowedUrls?: Array<AllowedUrlObject>
  /**
   * API definition identifier
   */
  apiId: string
  limits?: LimitObject
  quotas?: QuotaObject
  /**
   * Available versions list
   */
  versions?: Array<string>
}